Moved code around + bugfix
Fixed a bug with dispatch order where successful links were redirected to the fail page.

Ensured that last_updated_by is set when an account is linked.

Moved some code to the utils module.
